Create lab3.cpp
Не реализовано кодирование и декодирование сообщений
